requirements.txt
altgraph==0.17.4
biopython==1.84
cairocffi==1.7.1
cffi==1.17.1
chardet==5.2.0
cssselect2==0.7.0
defusedxml==0.7.1
et-xmlfile==1.1.0
lxml==5.3.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==10.4.0
pycparser==2.22
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
setuptools==75.1.0
six==1.16.0
tinycss2==1.3.0
tzdata==2024.2
webencodings==0.5.1
